@@ -1,6 +1,6 @@
 {
   "name": "strapi-helper-plugin",
-  "version": "3.0.0-alpha.14.1.1",
+  "version": "3.0.0-alpha.14.5",
   "description": "Helper for Strapi plugins development",
   "engines": {
     "node": ">= 9.0.0",
@@ -30,7 +30,6 @@
     "add-asset-html-webpack-plugin": "^2.1.2",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.0",
-    "babel-eslint": "^7.2.3",
     "babel-loader": "^7.1.1",
     "babel-plugin-istanbul": "^4.1.5",
     "babel-plugin-styled-components": "^1.5.1",
@@ -39,7 +38,7 @@
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-plugin-transform-react-constant-elements": "^6.23.0",
     "babel-plugin-transform-react-inline-elements": "^6.22.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.8",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.18",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
@@ -51,15 +50,6 @@
     "copy-webpack-plugin": "^4.3.1",
     "cross-env": "^5.0.5",
     "css-loader": "^0.28.5",
-    "eslint": "4.4.1",
-    "eslint-config-airbnb": "15.1.0",
-    "eslint-config-airbnb-base": "11.3.1",
-    "eslint-config-prettier": "2.3.0",
-    "eslint-import-resolver-webpack": "0.8.3",
-    "eslint-plugin-import": "2.8.0",
-    "eslint-plugin-jsx-a11y": "6.0.2",
-    "eslint-plugin-react": "7.2.1",
-    "eslint-plugin-redux-saga": "0.4.0",
     "exports-loader": "^0.6.4",
     "express": "^4.15.4",
     "extract-text-webpack-plugin": "^3.0.0",
@@ -67,15 +57,14 @@
     "history": "^4.6.3",
     "html-loader": "^0.5.1",
     "html-webpack-plugin": "^2.30.1",
-    "image-webpack-loader": "^3.3.1",
-    "immutable": "^3.8.1",
+    "image-webpack-loader": "^4.3.1",
+    "immutable": "^3.8.2",
     "imports-loader": "^0.7.1",
     "intl": "^1.2.5",
     "invariant": "2.2.1",
     "json-loader": "^0.5.7",
     "lodash": "^4.17.5",
     "lodash-webpack-plugin": "^0.11.4",
-    "mocha": "3.1.2",
     "moment": "^2.16.0",
     "node-sass": "^4.5.3",
     "null-loader": "^0.1.1",
@@ -88,29 +77,29 @@
     "precss": "^2.0.0",
     "prettier": "^1.5.3",
     "prop-types": "^15.5.10",
-    "react": "^16.0.0",
-    "react-datetime": "^2.8.6",
+    "react": "^16.5.2",
+    "react-datetime": "^2.15.0",
     "react-dnd": "^5.0.0",
     "react-dnd-html5-backend": "^5.0.1",
-    "react-dom": "^16.0.0",
+    "react-dom": "^16.5.2",
     "react-helmet": "^5.1.3",
-    "react-intl": "^2.3.0",
-    "react-loadable": "^5.3.1",
-    "react-redux": "^5.0.6",
-    "react-router-dom": "^4.1.2",
-    "react-router-redux": "^5.0.0-alpha.6",
-    "react-transition-group": "^2.2.0",
-    "reactstrap": "^5.0.0-alpha.3",
-    "redux": "^3.7.2",
+    "react-intl": "^2.7.0",
+    "react-loadable": "^5.5.0",
+    "react-redux": "^5.0.7",
+    "react-router-dom": "^4.3.1",
+    "react-router-redux": "^5.0.0-alpha.9",
+    "react-transition-group": "^2.5.0",
+    "reactstrap": "^5.0.0",
+    "redux": "^4.0.0",
     "redux-immutable": "^4.0.0",
-    "redux-saga": "^0.15.6",
+    "redux-saga": "^0.16.0",
     "reselect": "^3.0.1",
     "rimraf": "^2.6.1",
     "sass-loader": "^6.0.6",
     "shelljs": "^0.7.8",
     "style-loader": "^0.18.2",
-    "styled-components": "3.2.6",
-    "url-loader": "^0.5.9",
+    "styled-components": "^3.2.6",
+    "url-loader": "^1.1.1",
     "webpack": "^3.5.5",
     "webpack-bundle-analyzer": "^2.9.0",
     "webpack-dev-middleware": "^1.12.0",
