@@ -102,12 +102,10 @@ module.exports = async cb => {
         disabled: false,
       };
     });
-    
+
     // Don't display fields that are hidden by default like the resetPasswordToken for the model user
-    fieldsToRemove.forEach(field => {
-      _.unset(fields, field);
-      _.unset(schemaModel.attributes, field);
-    });
+    _.unset(fields, fieldsToRemove);
+    schemaModel.attributes = _.omit(schemaModel.attributes, fieldsToRemove);
 
     schemaModel.fields = fields;
     schemaModel.editDisplay.availableFields = fields;
@@ -293,6 +291,15 @@ module.exports = async cb => {
       // Update the displayed fields
       const updatedListDisplay = prevListDisplay.filter(obj => obj.name !== currentAttr.join());
 
+      // Retrieve the model's displayed fields for the `EditPage`
+      const fieldsPath = getEditDisplayFieldsPath(attrPath);
+      // Retrieve the previous settings
+      const prevEditDisplayFields = _.get(prevSchema.models, fieldsPath);
+      // Update the fields
+      const updatedEditDisplayFields = prevEditDisplayFields.filter(field => field !== currentAttr.join());
+      // Set the new layout
+      _.set(prevSchema.models, fieldsPath, updatedEditDisplayFields);
+
       if (updatedListDisplay.length === 0) {
         // Update it with the one from the generated schema
         _.set(prevSchema.models, listDisplayPath, _.get(schema.models, listDisplayPath, []));
@@ -332,16 +339,16 @@ module.exports = async cb => {
     });
 
     // Update other keys
-    sameApis.map(apiPath => {
+    sameApis.forEach(apiPath => {
       // This doesn't keep the prevSettings for the relations,  the user will have to reset it.
       // We might have to improve this if we want the order of the relations to be kept
-      const keysToUpdate = ['relations', 'loadedModel', 'associations', 'attributes', ['editDisplay', 'relations']].map(key => apiPath.concat(key));
+      ['relations', 'loadedModel', 'associations', 'attributes', ['editDisplay', 'relations']]
+        .map(key => apiPath.concat(key))
+        .forEach(keyPath => {
+          const newValue = _.get(schema.models, keyPath);
 
-      keysToUpdate.map(keyPath => {
-        const newValue = _.get(schema.models, keyPath);
-
-        _.set(prevSchema.models, keyPath, newValue);
-      });
+          _.set(prevSchema.models, keyPath, newValue);
+        });
     });
 
     // Special handler for the upload relations
@@ -356,6 +363,7 @@ module.exports = async cb => {
     });
 
     await pluginStore.set({ key: 'schema', value: prevSchema });
+
   } catch(err) {
     console.log('error', err);
   }
