@@ -1,7 +1,8 @@
 const _ = require('lodash');
+
 /**
  * Retrieve the path of each API
- * @param {Object}} data 
+ * @param {Object}} data
  * @returns {Array} Array of API path ['plugins.upload.file', 'plugins.users-permissions.user', ...]
  */
 const getApis = (data) => Object.keys(data).reduce((acc, curr) => {
@@ -10,8 +11,8 @@ const getApis = (data) => Object.keys(data).reduce((acc, curr) => {
   }
 
   if (curr === 'plugins') {
-    Object.keys(data[curr]).map(plugin => {
-      Object.keys(data[curr][plugin]).map(api => {
+    Object.keys(data[curr]).forEach(plugin => {
+      Object.keys(data[curr][plugin]).forEach(api => {
         acc = acc.concat([`${curr}.${plugin}.${api}`]);
       });
     });
@@ -23,8 +24,8 @@ const getApis = (data) => Object.keys(data).reduce((acc, curr) => {
 
 /**
  * Retrieve all the fields from an api
- * @param {Object} data 
- * @param {Array} apis 
+ * @param {Object} data
+ * @param {Array} apis
  * @returns {Array} Array composed of fields path for instance : [['plugins.users-permissions.user.fields.username', 'plugins.users-permissions.user.fields.email', 'plugins.users-permissions.user.fields.password'], [...]]
  */
 const getApisKeys = (data, apis) => apis.map(apiPath => {
@@ -43,12 +44,12 @@ const getApisUploadRelations = (data, sameArray) => sameArray.map(apiPath => {
     .filter(relationName => {
       return _.get(data.models, [...relationPath, relationName, 'plugin' ]) === 'upload';
     });
-  
+
   return relations.map(relation => `${apiPath.join('.')}.editDisplay.availableFields.${relation}`);
 });
 
 /**
- * 
+ *
  * @param {String} attrPath
  * @returns {Array}
  */
