@@ -13,7 +13,8 @@ const toArray = require('stream-to-array');
 const uuid = require('uuid/v4');
 
 function niceHash(buffer) {
-  return crypto.createHash('sha256')
+  return crypto
+    .createHash('sha256')
     .update(buffer)
     .digest('base64')
     .replace(/=/g, '')
@@ -35,7 +36,7 @@ module.exports = {
       files.map(async stream => {
         const parts = await toArray(fs.createReadStream(stream.path));
         const buffers = parts.map(
-          part => _.isBuffer(part) ? part : Buffer.from(part)
+          part => (_.isBuffer(part) ? part : Buffer.from(part)),
         );
 
         const buffer = Buffer.concat(buffers);
@@ -44,21 +45,30 @@ module.exports = {
           name: stream.name,
           sha256: niceHash(buffer),
           hash: uuid().replace(/-/g, ''),
-          ext: stream.name.split('.').length > 1 ? `.${_.last(stream.name.split('.'))}` : '',
+          ext:
+            stream.name.split('.').length > 1
+              ? `.${_.last(stream.name.split('.'))}`
+              : '',
           buffer,
           mime: stream.type,
-          size: (stream.size / 1000).toFixed(2)
+          size: (stream.size / 1000).toFixed(2),
         };
-      })
+      }),
     );
   },
 
   upload: async (files, config) => {
     // Get upload provider settings to configure the provider to use.
-    const provider = _.find(strapi.plugins.upload.config.providers, { provider: config.provider });
+    const provider = _.find(strapi.plugins.upload.config.providers, {
+      provider: config.provider,
+    });
 
     if (!provider) {
-      throw new Error(`The provider package isn't installed. Please run \`npm install strapi-upload-${config.provider}\``);
+      throw new Error(
+        `The provider package isn't installed. Please run \`npm install strapi-provider-upload-${
+          config.provider
+        }\``,
+      );
     }
 
     const actions = provider.init(config);
@@ -74,16 +84,22 @@ module.exports = {
         file.provider = provider.provider;
 
         return await strapi.plugins['upload'].services.upload.add(file);
-      })
+      }),
     );
   },
 
-  add: async (values) => {
+  add: async values => {
     // Use Content Manager business logic to handle relation.
     if (strapi.plugins['content-manager']) {
-      return await strapi.plugins['content-manager'].services['contentmanager'].add({
-        model: 'file'
-      }, values, 'upload');
+      return await strapi.plugins['content-manager'].services[
+        'contentmanager'
+      ].add(
+        {
+          model: 'file',
+        },
+        values,
+        'upload',
+      );
     }
 
     return strapi.query('file', 'upload').create(values);
@@ -93,20 +109,26 @@ module.exports = {
     // Use Content Manager business logic to handle relation.
     if (strapi.plugins['content-manager']) {
       params.model = 'file';
-      params.id = (params._id || params.id);
+      params.id = params._id || params.id;
 
-      return await strapi.plugins['content-manager'].services['contentmanager'].edit(params, values, 'upload');
+      return await strapi.plugins['content-manager'].services[
+        'contentmanager'
+      ].edit(params, values, 'upload');
     }
 
     return strapi.query('file', 'upload').update(_.assign(params, values));
   },
 
-  fetch: (params) => {
-    return strapi.query('file', 'upload').findOne(_.pick(params, ['_id', 'id']));
+  fetch: params => {
+    return strapi
+      .query('file', 'upload')
+      .findOne(_.pick(params, ['_id', 'id']));
   },
 
-  fetchAll: (params) => {
-    return strapi.query('file', 'upload').find(strapi.utils.models.convertParams('file', params));
+  fetchAll: params => {
+    return strapi
+      .query('file', 'upload')
+      .find(strapi.utils.models.convertParams('file', params));
   },
 
   count: async () => {
@@ -114,12 +136,16 @@ module.exports = {
   },
 
   remove: async (params, config) => {
-    params.id = (params._id || params.id);
+    params.id = params._id || params.id;
 
     const file = await strapi.plugins['upload'].services.upload.fetch(params);
 
     // get upload provider settings to configure the provider to use
-    const provider = _.cloneDeep(_.find(strapi.plugins.upload.config.providers, {provider: config.provider}));
+    const provider = _.cloneDeep(
+      _.find(strapi.plugins.upload.config.providers, {
+        provider: config.provider,
+      }),
+    );
     _.assign(provider, config);
     const actions = provider.init(config);
 
@@ -132,48 +158,55 @@ module.exports = {
     if (strapi.plugins['content-manager']) {
       params.model = 'file';
 
-      await strapi.plugins['content-manager'].services['contentmanager'].delete(params, {source: 'upload'});
+      await strapi.plugins['content-manager'].services['contentmanager'].delete(
+        params,
+        { source: 'upload' },
+      );
     }
 
     return strapi.query('file', 'upload').delete(params);
   },
 
-  uploadToEntity: async function (params, files, source) {
+  uploadToEntity: async function(params, files, source) {
     // Retrieve provider settings from database.
-    const config = await strapi.store({
-      environment: strapi.config.environment,
-      type: 'plugin',
-      name: 'upload'
-    }).get({ key: 'provider' });
+    const config = await strapi
+      .store({
+        environment: strapi.config.environment,
+        type: 'plugin',
+        name: 'upload',
+      })
+      .get({ key: 'provider' });
 
-    const model = source && source !== 'content-manager' ?
-      strapi.plugins[source].models[params.model]:
-      strapi.models[params.model];
+    const model =
+      source && source !== 'content-manager'
+        ? strapi.plugins[source].models[params.model]
+        : strapi.models[params.model];
 
     // Asynchronous upload.
     await Promise.all(
-      Object.keys(files)
-        .map(async attribute => {
-          // Bufferize files per attribute.
-          const buffers = await this.bufferize(files[attribute]);
-          const enhancedFiles = buffers.map(file => {
-            const details = model.attributes[attribute];
+      Object.keys(files).map(async attribute => {
+        // Bufferize files per attribute.
+        const buffers = await this.bufferize(files[attribute]);
+        const enhancedFiles = buffers.map(file => {
+          const details = model.attributes[attribute];
 
-            // Add related information to be able to make
-            // the relationships later.
-            file[details.via] = [{
+          // Add related information to be able to make
+          // the relationships later.
+          file[details.via] = [
+            {
               refId: params.id,
               ref: params.model,
               source,
               field: attribute,
-            }];
+            },
+          ];
 
-            return file;
-          });
+          return file;
+        });
 
-          // Make upload async.
-          return this.upload(enhancedFiles, config);
-        })
+        // Make upload async.
+        return this.upload(enhancedFiles, config);
+      }),
     );
-  }
+  },
 };
