@@ -1,16 +1,19 @@
-import { get, includes, isArray, set, omit } from 'lodash';
-import { call, fork, takeLatest, put, select } from 'redux-saga/effects';
+import { get, includes, isArray, omit, set } from 'lodash';
+import { call, fork, put, select, takeLatest } from 'redux-saga/effects';
 import auth from 'utils/auth';
 import request from 'utils/request';
 
-import { makeSelectFormType, makeSelectModifiedData } from './selectors';
 import { hideLoginErrorsInput, submitError, submitSucceeded } from './actions';
 import { SUBMIT } from './constants';
+import { makeSelectFormType, makeSelectModifiedData } from './selectors';
 
 export function* submitForm(action) {
+
   try {
-    const formType = yield select(makeSelectFormType());
     const body = yield select(makeSelectModifiedData());
+    const formType = yield select(makeSelectFormType());
+    const isRegister = formType === 'register';
+
     let requestURL;
 
     switch (formType) {
@@ -33,12 +36,13 @@ export function* submitForm(action) {
 
     const response = yield call(request, requestURL, { method: 'POST', body: omit(body, 'news') });
 
-    if (response.jwt) {
+    if(get(response, 'user.role.name', '') === 'Administrator' || isRegister){
+      
       yield call(auth.setToken, response.jwt, body.rememberMe);
       yield call(auth.setUserInfo, response.user, body.rememberMe);
     }
 
-    if (formType === 'register') {
+    if (isRegister) {
       action.context.updatePlugin('users-permissions', 'hasAdminUser', true);
 
       if (body.news) {
@@ -82,7 +86,11 @@ export function* submitForm(action) {
           yield put(hideLoginErrorsInput(true));
           break;
         case 'reset-password':
-          formErrors = [{ name: 'password', errors: [{ id: 'users-permissions.Auth.form.error.password.matching' }] }];
+          if (errors[0].id === 'users-permissions.Auth.form.error.code.provide') {
+            strapi.notification.error(errors[0].id);
+          } else {
+            formErrors = [{ name: 'password', errors }];
+          }
           break;
         case 'register': {
           const target = includes(get(errors, ['0', 'id']), 'username') ? 'username' : 'email';
