@@ -115,13 +115,15 @@ module.exports = {
         return acc;
       }, {}));
 
-    const appControllers = Object.keys(strapi.api || {}).reduce((acc, key) => {
-      Object.keys(strapi.api[key].controllers).forEach((controller) => {
-        acc.controllers[controller] = generateActions(strapi.api[key].controllers[controller]);
-      });
+    const appControllers = Object.keys(strapi.api || {})
+      .filter(key => !!strapi.api[key].controllers)
+      .reduce((acc, key) => {
+        Object.keys(strapi.api[key].controllers).forEach((controller) => {
+          acc.controllers[controller] = generateActions(strapi.api[key].controllers[controller]);
+        });
 
-      return acc;
-    }, { controllers: {} });
+        return acc;
+      }, { controllers: {} });
 
     const pluginsPermissions = Object.keys(strapi.plugins).reduce((acc, key) => {
       const initialState = {
@@ -190,12 +192,21 @@ module.exports = {
     const routes = Object.keys(strapi.api || {}).reduce((acc, current) => {
       return acc.concat(_.get(strapi.api[current].config, 'routes', []));
     }, []);
+    const clonedPlugins = _.cloneDeep(strapi.plugins);
+    const pluginsRoutes = Object.keys(clonedPlugins || {}).reduce((acc, current) => {
+      const routes = _.get(clonedPlugins, [current, 'config', 'routes'], [])
+        .reduce((acc, curr) => {
+          const prefix = curr.config.prefix;
+          const path = prefix !== undefined ? `${prefix}${curr.path}` : `/${current}${curr.path}`;
+          _.set(curr, 'path', path);
 
-    const pluginsRoutes = Object.keys(strapi.plugins || {}).reduce((acc, current) => {
-      acc[current] = _.get(strapi.plugins[current].config, 'routes', []);
+          return acc.concat(curr);
+        }, []);
+
+      acc[current] = routes;
 
       return acc;
-    }, []);
+    }, {});
 
     return _.merge({ application: routes }, pluginsRoutes);
   },
@@ -210,7 +221,7 @@ module.exports = {
     // Aggregate first level actions.
     const appActions = Object.keys(strapi.api || {}).reduce((acc, api) => {
       Object.keys(_.get(strapi.api[api], 'controllers', {}))
-        .map(controller => {
+        .forEach(controller => {
           const actions = Object.keys(strapi.api[api].controllers[controller])
             .filter(action => _.isFunction(strapi.api[api].controllers[controller][action]))
             .map(action => `application.${controller}.${action.toLowerCase()}`);
@@ -224,7 +235,7 @@ module.exports = {
     // Aggregate plugins' actions.
     const pluginsActions = Object.keys(strapi.plugins).reduce((acc, plugin) => {
       Object.keys(strapi.plugins[plugin].controllers)
-        .map(controller => {
+        .forEach(controller => {
           const actions = Object.keys(strapi.plugins[plugin].controllers[controller])
             .filter(action => _.isFunction(strapi.plugins[plugin].controllers[controller][action]))
             .map(action => `${plugin}.${controller}.${action.toLowerCase()}`);
