@@ -13,7 +13,8 @@ module.exports.nested = function() {
     // Load root configurations.
     new Promise((resolve, reject) => {
       glob('./config/**/*.*(js|json)', {
-        cwd: this.config.appPath
+        cwd: this.config.appPath,
+        dot: true
       }, (err, files) => {
         if (err) {
           return reject(err);
@@ -367,7 +368,7 @@ const enableHookNestedDependencies = function (name, flattenHooksConfig, force =
         });
 
         return apiModelsUsed.length !== 0;
-      }) || 0; // Filter model with the right connector
+      }); // Filter model with the right connector
 
     flattenHooksConfig[name] = {
       enabled: force || modelsUsed.length > 0 // Will return false if there is no model, else true.
@@ -382,31 +383,6 @@ const enableHookNestedDependencies = function (name, flattenHooksConfig, force =
   }
 };
 
-/**
- * Allow dynamic config values through
- * the native ES6 template string function.
- */
-const regex = /\$\{[^()]*\}/g;
-const excludeConfigPaths = ['info.scripts'];
-const templateConfigurations = function (obj, configPath = '') {
-  // Allow values which looks like such as
-  // an ES6 literal string without parenthesis inside (aka function call).
-  // Exclude config with conflicting syntax (e.g. npm scripts).
-  return Object.keys(obj).reduce((acc, key) => {
-    if (isPlainObject(obj[key]) && !isString(obj[key])) {
-      acc[key] = templateConfigurations(obj[key], `${configPath}.${key}`);
-    } else if (isString(obj[key])
-      && !excludeConfigPaths.includes(configPath.substr(1))
-      && obj[key].match(regex) !== null) {
-      acc[key] = eval('`' + obj[key] + '`'); // eslint-disable-line prefer-template
-    } else {
-      acc[key] = obj[key];
-    }
-
-    return acc;
-  }, {});
-};
-    
 const isAdminInDevMode = function () {
   try {
     fs.accessSync(path.resolve(this.config.appPath, 'admin', 'admin', 'build', 'index.html'), fs.constants.R_OK | fs.constants.W_OK);
